Don't raise when ProcessLater or AlreadyProcessed
When a trigger is already processed and we try to process it again we can just asume everything went fine.
When we expect the trigger needs to be processed later we can just set the `process_after` time and leave the trigger
be. It will automatically be processed at the correct time